mcp_readpaper/requirements.txt
PyMuPDF==1.23.8
pandas==2.2.1
google-generativeai==0.3.2
python-dotenv==1.0.1
PyPDF2==3.0.1 